refactor: move Runner's Task class to its own file
